CreateProviderUser method at UserService was covered by unit test and VerificationService was covered by unit tests

+++ src/main/java/com/notic/controller/VerificationController.java
@@ -63,7 +63,7 @@ public class VerificationController {
             @RequestBody
             VerificationCodeRequestDto body
     ) {
-            long parsedCode = Long.parseLong(body.code());
+            int parsedCode = Integer.parseInt(body.code());
             verificationService.verifyCode(parsedCode);
             String response = "Email successfully verified";
             return ResponseEntity.ok(response);
